chore: move embedded transformer to destination_transformer
